+++ jgclark.Dashboard/src/types.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Types for Dashboard code
-// Last updated 2026-07-12 for v2.4.0.b49 by @jgclark
+// Last updated 2026-07-14 for v2.4.0.b50 by @jgclark
 //-----------------------------------------------------------------------------
 
 // Types for Settings
@@ -44,7 +44,6 @@ export type TDashboardSettings = {
   FFlag_UseTagCache?: boolean, // TODO: remove this in v2.5.0?
   FFlag_UseTagCacheAPIComparison?: boolean, // TODO: remove this in v2.5.0?
   FFlag_DynamicAddToAnywhere?: boolean, // Use new DynamicDialog-based add task dialog instead of QuickCapture plugin,
-  FFlag_Reminders?: boolean, // for v2.5.0 beta testing; TODO: remove this before v2.5.0 release
 
 
   /* SETTINGS THAT ARE CALCULATED AND PASSED BY THE PLUGIN ------------- */